+++ models/demos/deepseek_v3_d_p/tests/pcc/test_ttnn_dispatch_combine.py
@@ -314,7 +314,7 @@ def test_ttnn_dispatch_combine(
         dtype=ttnn.bfloat16,
     )
     tt_indices = ttnn.from_torch(
-        indices.to(torch.int16),
+        indices,
         mesh_mapper=mesh_mapper_dispatch_inputs,
         layout=ttnn.ROW_MAJOR_LAYOUT,
         device=mesh_device,
@@ -668,7 +668,7 @@ def test_ttnn_dispatch_combine_overflow(mesh_device, num_links, topology, overfl
         dtype=ttnn.bfloat16,
     )
     tt_indices = ttnn.from_torch(
-        indices.to(torch.int16),
+        indices,
         mesh_mapper=mesh_mapper_dispatch_inputs,
         layout=ttnn.ROW_MAJOR_LAYOUT,
         device=mesh_device,
